+++ crates/edge_detection/src/filter2d.rs
